Removed receiver ilumination and only include src ilumination if boolean is true

+++ lib/inversion.cpp
@@ -223,7 +223,6 @@ void InversionAcoustic2D<T>::runGrad() {
     std::shared_ptr<rockseis::Image2D<T>> vpgrad;
     std::shared_ptr<rockseis::Image2D<T>> rhograd;
     std::shared_ptr<rockseis::Image2D<T>> srcilum;
-    std::shared_ptr<rockseis::Image2D<T>> recilum;
     std::shared_ptr<rockseis::Data2D<T>> wavgrad;
 
     // Create a sort class
@@ -293,12 +292,10 @@ void InversionAcoustic2D<T>::runGrad() {
         rhograd = std::make_shared<rockseis::Image2D<T>>(Rhogradfile, gmodel, 1, 1);
         rhograd->createEmpty();
 
-        srcilum = std::make_shared<rockseis::Image2D<T>>(Srcilumfile, gmodel, 1, 1);
-        srcilum->createEmpty();
-
-        recilum = std::make_shared<rockseis::Image2D<T>>(Recilumfile, gmodel, 1, 1);
-        recilum->createEmpty();
-
+        if(this->srcilumset){
+            srcilum = std::make_shared<rockseis::Image2D<T>>(Srcilumfile, gmodel, 1, 1);
+            srcilum->createEmpty();
+        }
 
         for(long int i=0; i<ngathers; i++) {
             vpgrad->stackImage(Vpgradfile + "-" + std::to_string(i));
@@ -306,10 +303,10 @@ void InversionAcoustic2D<T>::runGrad() {
             rhograd->stackImage(Rhogradfile + "-" + std::to_string(i));
             remove_file(Rhogradfile + "-" + std::to_string(i));
 
-            srcilum->stackImage(Srcilumfile + "-" + std::to_string(i));
-            remove_file(Srcilumfile + "-" + std::to_string(i));
-            recilum->stackImage(Recilumfile + "-" + std::to_string(i));
-            remove_file(Recilumfile + "-" + std::to_string(i));
+            if(this->srcilumset){
+                srcilum->stackImage(Srcilumfile + "-" + std::to_string(i));
+                remove_file(Srcilumfile + "-" + std::to_string(i));
+            }
         }
 
 		//Clear work vector 
@@ -386,15 +383,14 @@ void InversionAcoustic2D<T>::runGrad() {
                 vpgrad = std::make_shared<rockseis::Image2D<T>>(Vpgradfile + "-" + std::to_string(work.id), lmodel, 1, 1);
                 rhograd = std::make_shared<rockseis::Image2D<T>>(Rhogradfile + "-" + std::to_string(work.id), lmodel, 1, 1);
 
-                srcilum = std::make_shared<rockseis::Image2D<T>>(Srcilumfile + "-" + std::to_string(work.id), lmodel, 1, 1);
-                recilum = std::make_shared<rockseis::Image2D<T>>(Recilumfile + "-" + std::to_string(work.id), lmodel, 1, 1);
+                if(this->srcilumset){
+                    srcilum = std::make_shared<rockseis::Image2D<T>>(Srcilumfile + "-" + std::to_string(work.id), lmodel, 1, 1);
+                    fwi->setSrcilum(srcilum);
+                }
 
                 // Setting up gradient objects in fwi class
                 fwi->setVpgrad(vpgrad);
                 fwi->setRhograd(rhograd);
-
-                fwi->setSrcilum(srcilum);
-                fwi->setRecilum(recilum);
 
                 wavgrad = std::make_shared<rockseis::Data2D<T>>(source->getNtrace(), source->getNt(), source->getDt(), 0.0);
                 wavgrad->setField(rockseis::PRESSURE);
@@ -434,8 +430,9 @@ void InversionAcoustic2D<T>::runGrad() {
                 wavgrad->putTrace(Wavgradfile, work.id);
 
                 // Output ilumination
-                srcilum->write();
-                recilum->write();
+                if(this->srcilumset){
+                    srcilum->write();
+                }
 
                 // Output misfit
                 Fmisfit->append(Misfitfile);
@@ -467,7 +464,6 @@ void InversionAcoustic2D<T>::runGrad() {
                 rhograd.reset();
                 wavgrad.reset();
                 srcilum.reset();
-                recilum.reset();
                 fwi.reset();
                 work.status = WORK_FINISHED;
 
@@ -974,38 +970,6 @@ void InversionAcoustic2D<T>::applySrcilum()
             if(ilumodel[i] != 0.0){
                 vp[i] = norm*vp[i]/ilumodel[i];
                 rho[i] = norm*rho[i]/ilumodel[i];
-            }
-        }
-        model->writeModel();
-    }
-}
-
-template<typename T>
-void InversionAcoustic2D<T>::applyRecilum()
-{
-    if(this->getRecilum()){
-        // Models
-        std::shared_ptr<rockseis::ModelAcoustic2D<T>> model;
-        model = std::make_shared<rockseis::ModelAcoustic2D<T>>(VPGRADFILE, RHOGRADFILE, 1 ,0);
-        // Ilumination
-        std::shared_ptr<rockseis::ModelAcoustic2D<T>> ilum (new rockseis::ModelAcoustic2D<T>(Recilumfile, Recilumfile, 1 ,0));
-
-        // Mute model and write
-        model->readModel();
-        ilum->readModel();
-        T *vp, *rho, *ilumodel;
-        vp = model->getVp(); 
-        rho = model->getR(); 
-        ilumodel = ilum->getVp(); 
-        int i;
-        int N;
-
-        N = (model->getGeom())->getNtot();
-        for(i=0; i< N; i++)
-        {
-            if(ilumodel[i] != 0.0){
-                vp[i] = vp[i]/ilumodel[i];
-                rho[i] = rho[i]/ilumodel[i];
             }
         }
         model->writeModel();
